requirements.txt
crewai==0.28.7
langchain==0.1.17
langchain-core==0.1.52
langchain-openai==0.1.7
streamlit==1.34.0
pillow==10.3.0
crewai_tools
python-dotenv==1.0.0
